requirements.txt
certifi==2018.1.18127.0.0.1
chardet==3.0.4
colorama==0.3.9
coverage==4.5.1
idna==2.6
nose==1.3.7
paho-mqtt==1.3.1
pg8000==1.11.0
psycopg2-binary==2.7.4
pynmea2==1.12.0
pyserial==3.4
python-dateutil==2.6.1
PyYAML==3.12
rednose==1.3.0
requests==2.18.4
six==1.11.0
termstyle==0.1.11
testing.common.database==2.0.3
testing.postgresql==1.3.0
urllib3==1.22